+++ examples/simple.rs
@@ -680,8 +680,7 @@ impl SimpleWindow {
         button5.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button5.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button5.on_click(|x, widgets| {
-            get_widget_by_name(widgets, "BoxInLayoutWidget3".to_string())
-                .set_config(
+            get_widget_by_name(widgets, "BoxInLayoutWidget3".to_string()).set_config(
                 CONFIG_MAIN_COLOR,
                 Config::Color([
                     (rand::random::<u8>() as f32 / 255.0),
@@ -792,8 +791,7 @@ impl SimpleWindow {
         button2.set_numeric(CONFIG_BORDER_WIDTH, 2);
         button2.set_color(CONFIG_BORDER_COLOR, [0.0, 0.0, 0.0, 1.0]);
         button2.on_click(|x, widgets| {
-            get_widget_by_name(widgets, "ProgressWidget".to_string())
-                .set_config(
+            get_widget_by_name(widgets, "ProgressWidget".to_string()).set_config(
                 CONFIG_SECONDARY_COLOR,
                 Config::Color([
                     (rand::random::<u8>() as f32 / 255.0),