+++ src/widget/toggle_button_widget.rs
@@ -37,7 +37,7 @@ pub struct ToggleButtonWidget {
     selected: bool,
     active: bool,
     widget_id: i32,
-    on_click: Option<Box<dyn FnMut(&mut ToggleButtonWidget, bool)>>,
+    on_click: Option<Box<dyn FnMut(&mut ToggleButtonWidget, bool, &Vec<WidgetContainer>)>>,
 }
 
 impl ToggleButtonWidget {
@@ -94,7 +94,7 @@ impl ToggleButtonWidget {
     /// widget.
     pub fn on_click<F>(&mut self, callback: F)
     where
-        F: FnMut(&mut ToggleButtonWidget, bool) + 'static,
+        F: FnMut(&mut ToggleButtonWidget, bool, &Vec<WidgetContainer>) + 'static,
     {
         self.on_click = Some(Box::new(callback));
     }
@@ -103,9 +103,9 @@ impl ToggleButtonWidget {
     /// of the current `Widget` object as a parameter, so this object can be modified when
     /// a click is registered, if necessary.  Also indicates the state of the object, whether
     /// or not the object has been toggled.
-    pub fn click(&mut self, state: bool) {
+    pub fn click(&mut self, state: bool, widgets: &Vec<WidgetContainer>) {
         if let Some(mut cb) = self.on_click.take() {
-            cb(self, state);
+            cb(self, state, widgets);
             self.on_click = Some(cb);
         }
     }
@@ -149,7 +149,7 @@ impl Widget for ToggleButtonWidget {
         &mut self,
         injected: bool,
         event: CallbackEvent,
-        _widget_store: Option<&Vec<WidgetContainer>>,
+        widget_store: Option<&Vec<WidgetContainer>>,
     ) -> Option<CallbackEvent> {
         if !injected {
             match event {
@@ -184,7 +184,13 @@ impl Widget for ToggleButtonWidget {
                             self.selected = !self.selected;
                             self.draw_unhovered();
                             self.active = false;
-                            self.click(self.selected);
+
+                            match widget_store {
+                                Some(widgets) => {
+                                    self.click(self.selected, widgets);
+                                }
+                                None => (),
+                            }
 
                             return Some(WidgetSelected {
                                 widget_id,