+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## 0.4.5
 
 - Changed get_widget_position_by_name to get_widget_by_name (far easier to use)
+- Fixed all other on_click callbacks so that widgets from widget store are included
 
 ## 0.4.4